requirements.txt
asgiref
daphne

beautifulsoup4
django
django-ninja
pydantic[email]
psycopg[binary]
python-dotenv
faststream[redis, cli]
whitenoise